Rewritten code for discover, additional tweaks.
